+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinHealth/CgibinHealthGetReportJobInfoResponse.cs
@@ -25,7 +25,7 @@
                         /// </summary>
                         [Newtonsoft.Json.JsonProperty("partyids")]
                         [System.Text.Json.Serialization.JsonPropertyName("partyids")]
-                        public int[]? DepartmentIdList { get; set; }
+                        public long[]? DepartmentIdList { get; set; }
                     }
 
                     public class ReportTo
