+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinSchool/Department/CgibinSchoolDepartmentCreateRequest.cs
@@ -39,7 +39,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
         /// </summary>
         [Newtonsoft.Json.JsonProperty("id")]
         [System.Text.Json.Serialization.JsonPropertyName("id")]
-        public int? DepartmentId { get; set; }
+        public long? DepartmentId { get; set; }
 
         /// <summary>
         /// 获取或设置部门名称。
@@ -60,7 +60,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
         /// </summary>
         [Newtonsoft.Json.JsonProperty("parentid")]
         [System.Text.Json.Serialization.JsonPropertyName("parentid")]
-        public int ParentDepartmentId { get; set; }
+        public long ParentDepartmentId { get; set; }
 
         /// <summary>
         /// 获取或设置标准年级。
