+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinSchool/Department/CgibinSchoolDepartmentListResponse.cs
@@ -41,7 +41,7 @@
                 /// </summary>
                 [Newtonsoft.Json.JsonProperty("id")]
                 [System.Text.Json.Serialization.JsonPropertyName("id")]
-                public int DepartmentId { get; set; }
+                public long DepartmentId { get; set; }
 
                 /// <summary>
                 /// 获取或设置部门名称。
@@ -62,7 +62,7 @@
                 /// </summary>
                 [Newtonsoft.Json.JsonProperty("parentid")]
                 [System.Text.Json.Serialization.JsonPropertyName("parentid")]
-                public int ParentDepartmentId { get; set; }
+                public long ParentDepartmentId { get; set; }
 
                 /// <summary>
                 /// 获取或设置标准年级。
