+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinSchool/User/Student/CgibinSchoolUserUpdateStudentRequest.cs
@@ -33,6 +33,6 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
         /// </summary>
         [Newtonsoft.Json.JsonProperty("department")]
         [System.Text.Json.Serialization.JsonPropertyName("department")]
-        public IList<int>? DepartmentIdList { get; set; }
+        public IList<long>? DepartmentIdList { get; set; }
     }
 }
